09-01-Robotium/RobotiumSample/src/jp/mydns/sys1yagi/android/robotiumsample/RobotiumSampleActivity.java
package jp.mydns.sys1yagi.android.robotiumsample;

import android.app.Activity;
import android.content.Intent;
import android.os.Bundle;
import android.view.Menu;
import android.view.MenuItem;
import android.view.View;
import android.view.View.OnClickListener;
import android.widget.EditText;

public class RobotiumSampleActivity extends Activity {

    @Override
    protected void onCreate(Bundle savedInstanceState) {
        super.onCreate(savedInstanceState);
        setContentView(R.layout.activity_robotium_sample);

        findViewById(R.id.button).setOnClickListener(new OnClickListener() {
            @Override
            public void onClick(View v) {
                EditText editText = (EditText) findViewById(R.id.edit);
                if (!"".equals(editText.getText().toString())) {
                    startEditedActivity(editText.getText().toString());
                }
            }
        });
    }

    private void startEditedActivity(String text){
    	Intent intent = new Intent(this, EditedActivity.class);
        intent.putExtra(EditedActivity.EXTRAS_TEXT, text.toString());
        startActivity(intent);
    }
    
    @Override
    public boolean onCreateOptionsMenu(Menu menu) {
        getMenuInflater().inflate(R.menu.robotium_sample, menu);
        return true;
    }

    @Override
    public boolean onOptionsItemSelected(MenuItem item) {
        Intent intent = new Intent(this, SettingsActivity.class);
        startActivity(intent);
        return true;
    }

}
